+++ src/routes/distributors/index.ts
@@ -1,1 +1,4 @@
+export * from "./new";
 export * from "./list";
+export * from "./edit";
+export * from "./delete";